Implemented simple rate control proof of concept

+++ source/eu/fivegex/monitoring/control/udp/UDPControlPlaneXDRProducer.java
@@ -432,18 +432,18 @@ public class UDPControlPlaneXDRProducer extends AbstractUDPControlPlaneProducer 
     /* DC Control Service methods */
     
     @Override
-    public float getDCMeasurementsRate(ID dcId) throws ControlServiceException {
+    public Long getDCMeasurementsRate(ID dcId) throws ControlServiceException {
         List<Object> args = new ArrayList();
         args.add(dcId);
         
-        Float rate;
+        Long rate;
         
         ControlPlaneMessage m=new ControlPlaneMessage(ControlOperation.GET_DC_RATE, args);
         
         try {
             SocketControlEndPointMetaData dstAddr = (SocketControlEndPointMetaData)infoPlaneDelegate.getDCAddressFromID(dcId);
             MetaData mData = new UDPControlMetaData(dstAddr.getHost(), dstAddr.getPort());            
-            rate = (Float) synchronousTransmit(m, mData);
+            rate = (Long) synchronousTransmit(m, mData);
         } catch (IOException | DCNotFoundException | ControlPlaneConsumerException ex) {
             LOGGER.error("Error while performing getDCMeasurementsRate command " + ex.getMessage());
             throw new ControlServiceException(ex);
